strcasecmp compatability thing configure
