+++ src/audio/openal/al_audio_system.c
@@ -12,7 +12,7 @@
 
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 #include "stb_ds.h"
 
 // ===[ Helpers ]===
