+++ src/audio/ps2/ps2_audio_system.c
@@ -4,7 +4,7 @@
 
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 #include <inttypes.h>
 #include <audsrv.h>
