+++ src/data_win.c
@@ -4,7 +4,7 @@
 #include <stdbool.h>
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 #include "math_compat.h"
 
 #include "stb_ds.h"
