+++ src/desktop/backends/sdl1.c
@@ -1,5 +1,5 @@
 #include <ctype.h>
-#include <string.h>
+#include "string_compat.h"
 #include "stdio_compat.h"
 #include <ctype.h>
 #include <stdlib.h>
