+++ src/event_table.c
@@ -2,7 +2,7 @@
 #include "utils.h"
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 
 // ===[ EventSlotMap ]===
