+++ src/gl/gl_renderer.c
@@ -9,7 +9,7 @@
 #endif
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 #include "math_compat.h"
 
 #include "stb_image.h"
