+++ src/image/image_decoder.c
@@ -2,7 +2,7 @@
 
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 #include <bzlib.h>
 
 #include "stb_image.h"
