+++ src/ini.c
@@ -3,7 +3,7 @@
 
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 
 #include "text_utils.h"
