+++ src/input_recording.c
@@ -6,7 +6,7 @@
 
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 
 #include "stb_ds.h"
