+++ src/instance.c
@@ -1,7 +1,7 @@
 #include "instance.h"
 
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 #include "math_compat.h"
 
 #include "stb_ds.h"
