+++ src/noop_file_system.c
@@ -2,7 +2,7 @@
 #include "utils.h"
 
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 
 #include "stb_ds.h"
