+++ src/overlay_file_system.c
@@ -4,7 +4,7 @@
 
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 #include <sys/stat.h>
 #include <errno.h>
