+++ src/ps2/debug_font_renderer.c
@@ -10,7 +10,7 @@
 #include <malloc.h>
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 
 #define DEBUGFONT_LINE_HEIGHT_SCALE 0.80f
