+++ src/ps2/ps2_file_system.c
@@ -5,7 +5,7 @@
 
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 #include <sys/stat.h>
 
 #include "stb_ds.h"
