+++ src/ps3/main.c
@@ -5,7 +5,7 @@
 
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 #include <time.h>
 #include <malloc.h>
