+++ src/ps3/ps3_overlay.c
@@ -8,7 +8,7 @@
 #include <ps3gl.h>
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 
 #define OVERLAY_LINE_HEIGHT_SCALE 0.80f
 #define PROFILER_WINDOW_FRAMES 60
