+++ src/ps3/ps3_textures.c
@@ -2,7 +2,7 @@
 
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 
 // We stream the texture pages on demand from the file instead of loading everything in RAM.
