+++ src/runner.c
@@ -10,7 +10,7 @@
 #include <stdint.h>
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 #include "math_compat.h"
 
 #include "debug_overlay.h"
