+++ src/runner_gamepad.c
@@ -2,7 +2,7 @@
 #include "utils.h"
 
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 #include "math_compat.h"
 
 //DELTARUNE HACK
