+++ src/rvalue.h
@@ -4,7 +4,7 @@
 #include "common.h"
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 
 #include "real_type.h"
 #include "stb_ds.h"
