+++ src/string_builder.c
@@ -3,7 +3,7 @@
 
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 #include <stdarg.h>
 
 #define STRING_BUILDER_MIN_CAPACITY 16
