+++ src/vm_builtins.c
@@ -12,7 +12,7 @@
 
 #include "stdio_compat.h"
 #include <stdlib.h>
-#include <string.h>
+#include "string_compat.h"
 #include "math_compat.h"
 #include <ctype.h>
 #include <time.h>
